services/backend/pyproject.toml
[project]
name = "milio-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.10.3",
  "python-multipart==0.0.19",
  "sqlalchemy==2.0.36",
  "psycopg[binary]==3.2.3",
  "alembic==1.14.0",
  "boto3==1.35.70",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "cryptography==44.0.0",
  "openai-whisper",
  # Auth dependencies
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  # Rate limiting
  "slowapi==0.1.9",
]
